Created a version 2 of the snake game.
The snake model inherits the Turtle class.  It also separates the head(super) and tail.  The tail is now stored in a dictionary
called tail_segments.  Segments get pre-made or resued and are placed in a tail pool.
Based on the time lapsed in each game cycle, there is not much difference between v1 and
v2.  I may re-visit this one.